requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
PyJWT==2.4.0